Add forward declarations for units.h, remove it from creature.h (#42701)
* Add forward declarations for units.h

Should allow removing units.h from some headers, and speeding up build
times.

* Remove units.h from a few headers

Hopefully this has some impact.
