+++ src/item_contents.h
@@ -15,7 +15,7 @@
 #include "optional.h"
 #include "ret_val.h"
 #include "type_id.h"
-#include "units.h"
+#include "units_fwd.h"
 #include "visitable.h"
 
 class Character;
